+++ R/annotate_axis_line.R
@@ -19,7 +19,6 @@
 #' library(palmerpenguins)
 #'
 #' set_blanket()
-#' set_geom_reference_line()
 #'
 #' d <- penguins |>
 #'   add_row(