+++ R/gg_label.R
@@ -12,7 +12,6 @@
 #' library(dplyr)
 #'
 #' set_blanket()
-#' set_geom_font()
 #'
 #' bind_rows(
 #'   mtcars |> slice_min(order_by = mpg),